+++ Mukku/Mukku/MukkuApp.swift
@@ -1,0 +1,17 @@
+//
+//  MukkuApp.swift
+//  Mukku
+//
+//  Created by Kihyun Roh on 2023/05/04.
+//
+
+import SwiftUI
+
+@main
+struct MukkuApp: App {
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+        }
+    }
+}
