--- WidgetAgreed/MusicalWidget/MusicalWidgetLiveActivity.swift
@@ -1,86 +1,0 @@
-//
-//  MusicalWidgetLiveActivity.swift
-//  MusicalWidget
-//
-//  Created by polyappledev on 2023/05/01.
-//
-
-import ActivityKit
-import WidgetKit
-import SwiftUI
-
-struct MusicalWidgetAttributes: ActivityAttributes {
-    public struct ContentState: Codable, Hashable {
-        // Dynamic stateful properties about your activity go here!
-        var value: Int
-    }
-
-    // Fixed non-changing properties about your activity go here!
-    var name: String
-}
-
-
-struct MusicalWidgetLiveActivity: Widget {
-    var body: some WidgetConfiguration {
-        ActivityConfiguration(for: MusicalWidgetAttributes.self) { context in
-            // Lock screen/banner UI goes here
-            VStack {
-                Text("Hello")
-            }
-            .activityBackgroundTint(Color.cyan)
-            .activitySystemActionForegroundColor(Color.black)
-
-        } dynamicIsland: { context in
-            DynamicIsland {
-                // Expanded UI goes here.  Compose the expanded UI through
-                // various regions, like leading/trailing/center/bottom
-                DynamicIslandExpandedRegion(.leading) {
-                    Image("mask1")
-                        .resizable()
-                        .frame(width: 100, height: 100)
-                        .scaledToFit()
-//                    Text("Leading")
-                }
-                DynamicIslandExpandedRegion(.trailing) {
-                    Text("Trailing")
-                }
-                DynamicIslandExpandedRegion(.bottom) {
-                    Text("Bottom")
-                    // more content
-                }
-            } compactLeading: {
-                Image("mask1")
-                    .resizable()
-                    .frame(width: 30, height: 30)
-                    .scaledToFit()
-//                Text("L")
-            } compactTrailing: {
-                Text("T")
-            } minimal: {
-                Text("Min")
-            }
-            .widgetURL(URL(string: "http://www.apple.com"))
-            .keylineTint(Color.red)
-        }
-    }
-}
-
-struct MusicalWidgetLiveActivity_Previews: PreviewProvider {
-    static let attributes = MusicalWidgetAttributes(name: "Me")
-    static let contentState = MusicalWidgetAttributes.ContentState(value: 3)
-
-    static var previews: some View {
-        attributes
-            .previewContext(contentState, viewKind: .dynamicIsland(.compact))
-            .previewDisplayName("Island Compact")
-        attributes
-            .previewContext(contentState, viewKind: .dynamicIsland(.expanded))
-            .previewDisplayName("Island Expanded")
-        attributes
-            .previewContext(contentState, viewKind: .dynamicIsland(.minimal))
-            .previewDisplayName("Minimal")
-        attributes
-            .previewContext(contentState, viewKind: .content)
-            .previewDisplayName("Notification")
-    }
-}
