--- WidgetAgreed/WidgetAgreed/WidgetAgreedApp.swift
@@ -1,18 +1,0 @@
-//
-//  WidgetAgreedApp.swift
-//  WidgetAgreed
-//
-//  Created by polyappledev on 2023/05/01.
-//
-
-import SwiftUI
-
-// testing edit
-@main
-struct WidgetAgreedApp: App {
-    var body: some Scene {
-        WindowGroup {
-            ContentView()
-        }
-    }
-}
